# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agenda Inteligente — agenda-service",
...
    "version": "0.4.0"
  },
  "servers": [
    {
      "url": "https://cpdf-agenda.better-knowledge.com",
      "description": "Protótipo do programa"
    }
  ],
  "paths": {
...
    "/appointments": {
      "post": {
        "tags": [
          "agendamentos"
        ],
        "summary": "Agenda um horário",
        "description": "**Escopo:** `agenda:write`\n\nExige serviço + horário + cliente identificado (nome e telefone) — RF-03. Sem resource_id, usa o primeiro recurso do serviço com o slot livre. Horário ocupado responde 409 SLOT_INDISPONIVEL com as 3 alternativas mais próximas já no payload. Aceita Idempotency-Key (obrigatório para agentes).",
        "operationId": "agendar_appointments_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentOut"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "examples": {
                  "NAO_AUTENTICADO": {
                    "summary": "Credencial ausente ou inválida",
                    "value": {
                      "code": "NAO_AUTENTICADO",
                      "message": "Credencial ausente ou inválida: nenhum header de credencial presente",
                      "hint": "Envie `Authorization: Bearer <jwt do Supabase>` ou `X-Agent-Key`.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "403": {
            "description": "Escopo insuficiente",
            "content": {
              "application/json": {
                "examples": {
                  "ESCOPO_INSUFICIENTE": {
                    "summary": "A credencial não tem o escopo exigido",
                    "value": {
                      "code": "ESCOPO_INSUFICIENTE",
                      "message": "A credencial não tem o escopo 'agenda:cancel'.",
                      "hint": "Peça uma credencial com o escopo necessário — cancelamento exige agenda:cancel.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "422": {
            "description": "Validação falhou",
            "content": {
              "application/json": {
                "examples": {
                  "PAYLOAD_INVALIDO": {
                    "summary": "Corpo ou parâmetros não passaram na validação",
                    "value": {
                      "code": "PAYLOAD_INVALIDO",
                      "message": "Payload inválido: body.cliente_telefone: String should have at least 8 characters",
                      "hint": "Corrija os campos apontados e repita a chamada com os mesmos dados.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado nesta organização",
            "content": {
              "application/json": {
                "examples": {
                  "NAO_ENCONTRADO": {
                    "summary": "Id não existe nesta organização",
                    "value": {
                      "code": "NAO_ENCONTRADO",
                      "message": "Serviço 6f1e…9c02 não existe nesta organização.",
                      "hint": "Confira o id — liste o recurso correspondente para obter ids válidos.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflito — o hint diz como resolver",
            "content": {
              "application/json": {
                "examples": {
                  "SLOT_INDISPONIVEL": {
                    "summary": "Horário ocupado — alternativas já no payload",
                    "value": {
                      "code": "SLOT_INDISPONIVEL",
                      "message": "O horário pedido já está ocupado ou fora da grade.",
                      "hint": "Ofereça estas alternativas ao cliente: quinta, 27 de agosto, 15h30 · quinta, 27 de agosto, 16h30 · sexta, 28 de agosto, 9h.",
                      "retryable": false,
                      "alternativas": [
                        {
                          "inicio": "2026-08-27T15:30:00-03:00",
                          "label_humano": "quinta, 27 de agosto, 15h30"
                        },
                        {
                          "inicio": "2026-08-27T16:30:00-03:00",
                          "label_humano": "quinta, 27 de agosto, 16h30"
                        },
                        {
                          "inicio": "2026-08-28T09:00:00-03:00",
                          "label_humano": "sexta, 28 de agosto, 9h"
                        }
                      ]
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "400": {
            "description": "Pedido malformado",
            "content": {
              "application/json": {
                "examples": {
                  "DATA_SEM_FUSO": {
                    "summary": "Horário sem offset de fuso",
                    "value": {
                      "code": "DATA_SEM_FUSO",
                      "message": "'inicio' veio sem offset de fuso horário.",
                      "hint": "Envie ISO 8601 com offset explícito, ex.: 2026-08-27T15:30:00-03:00.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          }
        },
        "x-escopo-requerido": "agenda:write",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Repetir a chamada com a mesma chave devolve a resposta original sem duplicar efeito. Obrigatório para agentes; use um UUID novo por intenção."
          }
        ]
      },
      "get": {
        "tags": [
          "agendamentos"
        ],
        "summary": "Lista compromissos por data e recurso",
        "description": "**Escopo:** `agenda:operacao`\n\nTodos os status, na ordem do dia. Exige `agenda:operacao` — devolve nome, contato e observações de **todos** os clientes daquele dia. Para checar disponibilidade use GET /slots; para o compromisso de um cliente, GET /appointments/proximo.",
        "operationId": "listar_appointments_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "resource",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentOut"
                  },
                  "title": "Response Listar Appointments Get"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "examples": {
                  "NAO_AUTENTICADO": {
                    "summary": "Credencial ausente ou inválida",
                    "value": {
                      "code": "NAO_AUTENTICADO",
                      "message": "Credencial ausente ou inválida: nenhum header de credencial presente",
                      "hint": "Envie `Authorization: Bearer <jwt do Supabase>` ou `X-Agent-Key`.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "403": {
            "description": "Escopo insuficiente",
            "content": {
              "application/json": {
                "examples": {
                  "ESCOPO_INSUFICIENTE": {
                    "summary": "A credencial não tem o escopo exigido",
                    "value": {
                      "code": "ESCOPO_INSUFICIENTE",
                      "message": "A credencial não tem o escopo 'agenda:cancel'.",
                      "hint": "Peça uma credencial com o escopo necessário — cancelamento exige agenda:cancel.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "422": {
            "description": "Validação falhou",
            "content": {
              "application/json": {
                "examples": {
                  "PAYLOAD_INVALIDO": {
                    "summary": "Corpo ou parâmetros não passaram na validação",
                    "value": {
                      "code": "PAYLOAD_INVALIDO",
                      "message": "Payload inválido: body.cliente_telefone: String should have at least 8 characters",
                      "hint": "Corrija os campos apontados e repita a chamada com os mesmos dados.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          }
        },
        "x-escopo-requerido": "agenda:operacao"
      }
    },
...
    "/appointments/{appointment_id}/confirm": {
      "post": {
        "tags": [
          "agendamentos"
        ],
        "summary": "Marca o compromisso como confirmado pelo cliente",
        "description": "**Escopo:** `agenda:write`\n\nUse quando o cliente responder 'sim' à confirmação. Só compromissos 'agendado' aceitam.",
        "operationId": "confirmar_appointments__appointment_id__confirm_post",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentOut"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "examples": {
                  "NAO_AUTENTICADO": {
                    "summary": "Credencial ausente ou inválida",
                    "value": {
                      "code": "NAO_AUTENTICADO",
                      "message": "Credencial ausente ou inválida: nenhum header de credencial presente",
                      "hint": "Envie `Authorization: Bearer <jwt do Supabase>` ou `X-Agent-Key`.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "403": {
            "description": "Escopo insuficiente",
            "content": {
              "application/json": {
                "examples": {
                  "ESCOPO_INSUFICIENTE": {
                    "summary": "A credencial não tem o escopo exigido",
                    "value": {
                      "code": "ESCOPO_INSUFICIENTE",
                      "message": "A credencial não tem o escopo 'agenda:cancel'.",
                      "hint": "Peça uma credencial com o escopo necessário — cancelamento exige agenda:cancel.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "422": {
            "description": "Validação falhou",
            "content": {
              "application/json": {
                "examples": {
                  "PAYLOAD_INVALIDO": {
                    "summary": "Corpo ou parâmetros não passaram na validação",
                    "value": {
                      "code": "PAYLOAD_INVALIDO",
                      "message": "Payload inválido: body.cliente_telefone: String should have at least 8 characters",
                      "hint": "Corrija os campos apontados e repita a chamada com os mesmos dados.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado nesta organização",
            "content": {
              "application/json": {
                "examples": {
                  "NAO_ENCONTRADO": {
                    "summary": "Id não existe nesta organização",
                    "value": {
                      "code": "NAO_ENCONTRADO",
                      "message": "Serviço 6f1e…9c02 não existe nesta organização.",
                      "hint": "Confira o id — liste o recurso correspondente para obter ids válidos.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflito — o hint diz como resolver",
            "content": {
              "application/json": {
                "examples": {
                  "STATUS_INCOMPATIVEL": {
                    "summary": "O status atual não permite a ação",
                    "value": {
                      "code": "STATUS_INCOMPATIVEL",
                      "message": "Compromisso está 'cancelado' — não dá para reagendar.",
                      "hint": "Crie um novo agendamento com POST /appointments.",
                      "retryable": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErroOut"
                }
              }
            }
          }
        },
        "x-escopo-requerido": "agenda:write"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppointmentIn": {
        "properties": {
          "service_id": {
            "type": "string",
            "format": "uuid",
            "title": "Service Id"
          },
          "inicio": {
            "type": "string",
            "format": "date-time",
            "title": "Inicio",
            "description": "Um inicio devolvido por GET /slots (ISO 8601 com offset)"
          },
          "cliente_nome": {
            "type": "string",
            "minLength": 1,
            "title": "Cliente Nome"
          },
          "cliente_telefone": {
            "type": "string",
            "minLength": 3,
            "title": "Cliente Telefone",
            "description": "Endereço do cliente no canal: E.164 no WhatsApp (+5511998765432) ou tg:<chat_id> no Telegram. É por ele que os lembretes saem."
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "origem": {
            "type": "string",
            "title": "Origem",
            "description": "agente | cliente | humano | calendly",
            "default": "agente"
          },
          "observacoes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observacoes"
          }
        },
        "type": "object",
        "required": [
          "service_id",
          "inicio",
          "cliente_nome",
          "cliente_telefone"
        ],
        "title": "AppointmentIn",
        "examples": [
          {
            "cliente_nome": "Paula Andrade",
            "cliente_telefone": "+5511998765432",
            "inicio": "2026-08-27T15:30:00-03:00",
            "observacoes": "Prefere atendimento com a Júlia",
            "origem": "agente",
            "service_id": "b3f0a1d4-2c5e-4f6a-8b7c-9d0e1f2a3b4c"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "SupabaseJWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT do Supabase Auth com claim `org_id` — humanos e UI."
      },
      "AgentKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Agent-Key",
        "description": "Chave estática de agente (fase 1 do conector — vira OAuth 2.1 no agenda-mcp). Ações irreversíveis pedem confirmação humana."
      }
    }
  },
...
  "security": [
    {
      "SupabaseJWT": []
    },
    {
      "AgentKey": []
    }
  ]
}